+++ pgExtractLoadData.ts
@@ -241,9 +241,8 @@ export class pgExtractLoadData {
             for (let i = 0; i < filesDirSource1.length; i++) {
                 tableName = filesDirSource1[i].file;
                 tableName = tableName.substring(0, tableName.length - 4).toLowerCase(); //quito extension
-
-                await this.pgDb.query('ALTER TABLE ' + globalFunction.quotedString(tableName) + ' DISABLE TRIGGER ALL');
-
+                if (objectName === '' || objectName.toLowerCase() === tableName.toLowerCase()) 
+                   await this.pgDb.query('ALTER TABLE ' + globalFunction.quotedString(tableName) + ' DISABLE TRIGGER ALL');
             }
             //await this.pgDb.query('COMMIT');
 
@@ -254,9 +253,8 @@ export class pgExtractLoadData {
                 for (let i = 0; i < filesDirSource1.length; i++) {
                     tableName = filesDirSource1[i].file;
                     tableName = tableName.substring(0, tableName.length - 4).toLowerCase(); //quito extension
-
-                    await this.pgDb.query('DELETE FROM ' + globalFunction.quotedString(tableName));
-
+                    if (objectName === '' || objectName.toLowerCase() === tableName.toLowerCase()) 
+                        await this.pgDb.query('DELETE FROM ' + globalFunction.quotedString(tableName));
                 }
                 //await this.pgDb.query('COMMIT');
             }
@@ -264,15 +262,16 @@ export class pgExtractLoadData {
             for (let i = 0; i < filesDirSource1.length; i++) {
                 tableName = filesDirSource1[i].file;
                 tableName = tableName.substring(0, tableName.length - 4).toLowerCase(); //quito extension
+                if (objectName === '' || objectName.toLowerCase() === tableName.toLowerCase()) {
+                    console.log('Importando ' + tableName);
 
-                console.log('Importando ' + tableName);
+                    //await this.pgDb.query('BEGIN');
+                    stream = this.pgDb.query(copyFrom('COPY ' + globalFunction.quotedString(tableName) + ' FROM STDIN WITH CSV HEADER'));
+                    fileStream = fs.createReadStream(filesDirSource1[i].path + filesDirSource1[i].file)
+                    await execute(stream, fileStream);
 
-                //await this.pgDb.query('BEGIN');
-                stream = this.pgDb.query(copyFrom('COPY ' + globalFunction.quotedString(tableName) + ' FROM STDIN WITH CSV HEADER'));
-                fileStream = fs.createReadStream(filesDirSource1[i].path + filesDirSource1[i].file)
-                await execute(stream, fileStream);
-
-                // await this.pgDb.query('COMMIT');
+                    // await this.pgDb.query('COMMIT');
+                }    
             }
         }
         catch (err) {
@@ -285,8 +284,9 @@ export class pgExtractLoadData {
             for (let i = 0; i < filesDirSource1.length; i++) {
                 tableName = filesDirSource1[i].file;
                 tableName = tableName.substring(0, tableName.length - 4).toLowerCase(); //quito extension
-
-                await this.pgDb.query('ALTER TABLE ' + globalFunction.quotedString(tableName) + ' ENABLE TRIGGER ALL');
+                if (objectName === '' || objectName.toLowerCase() === tableName.toLowerCase()) { 
+                    await this.pgDb.query('ALTER TABLE ' + globalFunction.quotedString(tableName) + ' ENABLE TRIGGER ALL');
+                }    
             }
             //await this.pgDb.query('COMMIT');
         }
